requirements.txt
gradio>=4.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.1
supabase>=2.0.0
python-dotenv>=1.0.0
duckduckgo-search>=3.0.0
requests>=2.28.0
matplotlib>=3.5.0
pandas>=1.5.0
